+++ Source/ACE/Network/GameMessages/Messages/GameMessagePrivateUpdateVital.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using ACE.Network.Enum;
 using ACE.Entity;
+using ACE.Entity.Enum;
 
 namespace ACE.Network.GameMessages.Messages
 {